agent/python/requirements.txt
# ============================================
# Markov Audit System - Python Dependencies
# Version: 2.0.0
# Updated: 2025-10-26 05:46:53 UTC
# Developer: charlesms-eth
# License: MIT (Free Open Source)
# ============================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Agent Framework (Fetch.ai)
uagents==0.12.0

# MeTTa Reasoning Engine
hyperon==0.1.12

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# Report Generation
reportlab==4.0.7
markdown-it-py==3.0.0

# Contract Parsing
solc-select==1.0.4
py-solc-x==2.0.2

# Blockchain
web3==6.15.0
eth-utils==2.3.1

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0